requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coinbase==2.0.6
cytoolz==0.8.2
ethereum-abi-utils==0.4.0
ethereum-utils==0.4.0
future==0.16.0
idna==2.6
peewee==2.10.1
peewee-db-evolve==0.6.4
psycopg2==2.7.3.1
pycrypto==2.6.1
pylru==1.0.9
pysha3==1.0.2
python-telegram-bot==8.0
requests==2.18.4
rlp==0.6.0
six==1.11.0
toolz==0.8.2
urllib3==1.22
web3==3.13.5
